+++ local/shiny-apps/issue-029-r/app.R
@@ -1,0 +1,22 @@
+library(shiny)
+
+ui <- fluidPage(
+  selectInput("dataset", "Choose a dataset", c("pressure", "cars")),
+  selectInput("column", "Choose column", character(0)),
+  verbatimTextOutput("summary")
+)
+
+server <- function(input, output, session) {
+  dataset <- reactive(get(input$dataset, "package:datasets"))
+
+  observeEvent(input$dataset, {
+    freezeReactiveValue(input, "column")
+    updateSelectInput(inputId = "column", choices = names(dataset()))
+  })
+
+  output$summary <- renderPrint({
+    summary(dataset()[[input$column]])
+  })
+}
+
+shinyApp(ui, server)