+++ NEWS.md
@@ -1,5 +1,7 @@
 # shinylive 0.1.1
 
+* Bump shinylive assets dependency to 0.1.1.
+
 # shinylive 0.1.0
 
 * Initial CRAN submission.