+++ NEWS.md
@@ -1,5 +1,7 @@
 # shinylive (development version)
 
+* Use `{httpuv}` to serve static folder instead of plumber. (#40)
+
 # shinylive 0.1.0
 
 * Initial CRAN submission.