+++ NEWS.md
@@ -1,5 +1,7 @@
 # shinylive (development version)
 
+* Use `{httr2}` to download assets from GitHub releases. (@dgkf #30, #39)
+
 # shinylive 0.1.0
 
 * Initial CRAN submission.