+++ NEWS.md
@@ -1,5 +1,6 @@
 # shinylive (development version)
 
+* Use `{httpuv}` to serve static folder instead of plumber. (#40)
 * Use `{httr2}` to download assets from GitHub releases. (@dgkf #30, #39)
 
 # shinylive 0.1.0