+++ NEWS.md
@@ -1,7 +1,6 @@
 # shinylive 0.1.1
 
-* Bump shinylive assets dependency to 0.1.1.
-
+* Bump shinylive assets dependency to 0.2.3. (#38)
 * Use `{httpuv}` to serve static folder instead of plumber. (#40)
 * Use `{httr2}` to download assets from GitHub releases. (@dgkf #30, #39)